Split CrawlBar macOS surfaces
