+++ Sources/CrawlBar/Menu/CrawlBarMenuMetrics.swift
@@ -1,0 +1,12 @@
+import CoreGraphics
+
+enum CrawlBarMenuMetrics {
+    static let width: CGFloat = 380
+    static let proposedHeight: CGFloat = 720
+    static let maxHeight: CGFloat = 360
+    static let fallbackHeight: CGFloat = 28
+    static let selectionHorizontalInset: CGFloat = 6
+    static let selectionVerticalInset: CGFloat = 2
+    static let selectionCornerRadius: CGFloat = 6
+    static let submenuIndicatorTrailingPadding: CGFloat = 10
+}
