+++ Sources/CrawlBar/Settings/SettingsLayout.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+enum CrawlBarSettingsLayout {
+    static let minWindowWidth: CGFloat = 1120
+    static let minWindowHeight: CGFloat = 790
+    static let sidebarWidth: CGFloat = 250
+    static let detailHorizontalPadding: CGFloat = 22
+    static let detailVerticalPadding: CGFloat = 18
+    static let detailBottomPadding: CGFloat = 16
+}
